requirements.txt
flask
pyjwt
flask-redis
flask_sqlalchemy
psycopg2-binary
flask-cors